% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_harvest_rate}
\alias{adult_harvest_rate}
\title{Adult Harvest Rate}
\format{
1 dimensional array [31 watersheds]
}
\source{
Various \href{https://s3-us-west-2.amazonaws.com/cvpia-reference-docs/AWP+Attachment+1.pdf}{CVPIA Science Integration Team: FY18 Decision Support Model activities and FY17 priorities Memorandum}

Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
adult_harvest_rate
}
\description{
Proportion of adults harvested from golden gate until they reach their natal shed
}
\keyword{datasets}
