% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_seeds}
\alias{adult_seeds}
\title{Adult Seeds}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 31 rows and 30 columns.
}
\source{
Derived from average escapement estimates from 2013 to 2017 \href{https://dsm-docs.s3.us-west-2.amazonaws.com/Azat+2019.pdf}{Azat 2019}.
}
\usage{
adult_seeds
}
\description{
Adult fish for the initial 5 years of the simulations.
}
\details{
Adult Seeds
}
\keyword{datasets}
