% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{adult_stray}
\alias{adult_stray}
\title{Adult Straying}
\source{
IP-117068
}
\usage{
adult_stray(
  wild,
  natal_flow,
  south_delta_watershed,
  cross_channel_gates_closed,
  prop_delta_trans = 0,
  prop_bay_trans = 0,
  betas = c(intercept = 3, wild = -5.5, `natal flow` = -1.99, `gates open` = -0.174,
    bay = 2.09, delta = 2.89)
)
}
\arguments{
\item{wild}{Indicator of wild fish returning}

\item{natal_flow}{Proportion flows at tributary junctions coming from natal watershed in October}

\item{south_delta_watershed}{Indicator if watershed feeds into South Delta}

\item{cross_channel_gates_closed}{Number of days gates are closed for each month}

\item{betas}{Parameters estimated from calibration}
}
\description{
Calculate the proportion of adults straying to non-natal streams to spawn
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
or meta-analysis. The source for each parameter in this function are detailed below.


\itemize{
\item intercept- Intercept: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District
\item wild - Wild Salmon Parameter: Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}
\item natal flow - Natal Flow Parameter: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District
\item gates open - Cross Channel Gates Parameter: Empirical model fit using  2008–2011 tagging data provided by East Bay Municipal Utility District.
\item bay - Bay transport proportion parameter. Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}
\item delta  - Delta transport proportion parameter. Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}
}
}

