% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{apply_straying}
\alias{apply_straying}
\title{Apply Straying}
\usage{
apply_straying(
  year,
  natural_adults,
  hatchery_adults,
  total_releases,
  release_month,
  flows_oct_nov,
  flows_apr_may,
  monthly_mean_pdo
)
}
\arguments{
\item{total_releases}{= ..params$hatchery_release,}

\item{release_month}{= 1,}

\item{flows_oct_nov}{= ..params$flows_oct_nov,}

\item{flows_apr_may}{= ..params$flows_apr_may,}

\item{monthly_mean_pdo}{= fallRunDSM::monthly_mean_pdo,}
}
\description{
Apply Straying
}
