% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioenergetics_transitions}
\alias{bioenergetics_transitions}
\title{Bioenergetics Transistion Matrices}
\format{
4 x 4 x 28 x 8 array [4 size classes x 4 size classes x 28 temperatures x 8 rear/flood and level combinations]
}
\source{
TODO
}
\usage{
bioenergetics_transitions
}
\description{
Chinook growth rates in floodplain and inchannel habitat.
}
\keyword{datasets}
