% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_channel_stray_rate}
\alias{cross_channel_stray_rate}
\title{Cross Channel Stray Rate}
\format{
1 dimensional array [31 watersheds]
}
\source{
Expert opinion from SIT members.
}
\usage{
cross_channel_stray_rate
}
\description{
Natural straying allocation across channels.
}
\keyword{datasets}
