% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvpiaModels-package.R
\docType{package}
\name{cvpiaModels-package}
\alias{cvpiaModels}
\alias{cvpiaModels-package}
\title{cvpiaModels: Structured Decision Models for CVPIA}
\description{
SDM is a key effort by CVPIA agencies to continuously improve project 
    recommendations and prioritization toward the anadromous fish doubling goal. 
    Our intent is to improve collaboration under a science-based framework that is 
    both transparent and efficient. This package contains life cycle models for the 
    for the key species impacted by CVPIA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CVPIA-OSC.github.io/fallRunDSM/}
}

}
\author{
\strong{Maintainer}: Emanuel Rodriguez \email{erodriguez@flowwest.com}

Authors:
\itemize{
  \item James Petterson \email{jt.peterson@oregonstate.edu}
  \item Adam Duarte \email{adam.duarte@oregonstate.edu}
  \item Sadie Gill \email{sgill@flowwest.com}
}

}
