% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{days_in_month}
\alias{days_in_month}
\title{Days in Month}
\usage{
days_in_month(
  month,
  days_by_month = c(Jan = 31L, Feb = 28L, Mar = 31L, Apr = 30L, May = 31L, Jun = 30L,
    Jul = 31L, Aug = 31L, Sep = 30L, Oct = 31L, Nov = 30L, Dec = 31L)
)
}
\arguments{
\item{month}{Integer representation of month}
}
\description{
Number of days withing a month
}
