% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diversity_group}
\alias{diversity_group}
\title{Diversity Groups}
\format{
1 dimensional named array [31 watersheds]
}
\source{
The diversity groups are sourced from the ["Framework for Assessing Viability of Threatened and Endangered Chinook Salmon and Steelhead in the Sacramento-San Joaquin Basin"](http://escholarship.org/uc/item/3653x9xc) with slight modification.
The "Northern Sierra" diversity group covers a very large geographic area and was divided into two groups based on proximity.
}
\usage{
diversity_group
}
\description{
Watershed groups.
}
\keyword{datasets}
