% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fall_hatchery_release}
\alias{fall_hatchery_release}
\title{Fall Hatchery Release}
\format{
A matrix with dimension 31 x 4 (watershed x size class)
}
\source{
Hatchery return data comes from the production targets defined in the
\href{https://swfsc-publications.fisheries.noaa.gov/publications/CR/2012/2012California.pdf}{California HSRG (Pages 65 - 95)}
}
\usage{
fall_hatchery_release
}
\description{
Hatchery release by size class and hatchery
}
\keyword{datasets}
