% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed-and-sim.R
\name{fall_run_seed_and_sim}
\alias{fall_run_seed_and_sim}
\title{Seed and Simulate}
\usage{
fall_run_seed_and_sim(
  scenario = NULL,
  ..params = fallRunDSM::params,
  stochastic = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scenario}{a scenarion matrix to use}

\item{..params}{model params list to use for seeding and simulation}

\item{stochastic}{should the seedind and sim be run in stochastic mode?}
}
\description{
A wrapper function to run both seeding and simulation with one function call. Using this function
also ensures that both seed and sim are run with same scenario and params.
}
