% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fecundity_by_age}
\alias{fecundity_by_age}
\title{Fecundity Lookup}
\format{
a data frame with 7 rows and 3 columns (age, origin, fecundity)
}
\source{
\href{https://www.x-mol.net/paper/article/1664889742509719552}{Roni and Quinn 1995} logic was used to map age to size.
Then we scaled fecundity by size using logic from \href{https://reorienting-to-recovery.s3.us-west-1.amazonaws.com/RoniandQuinn1995.pdf}{Malick et all 2023}
}
\usage{
fecundity_by_age
}
\description{
a lookup table describing the fecundity of fish by age and origin
}
\keyword{datasets}
