% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_natal}
\alias{fill_natal}
\title{Filling Natal Tributary Habitat}
\source{
IP-117068
}
\usage{
fill_natal(
  juveniles,
  inchannel_habitat,
  floodplain_habitat,
  territory_size = c(0.0498944803729701, 0.138941944739835, 0.471083652829798, 0),
  up_to_size_class = 2
)
}
\arguments{
\item{juveniles}{An n by 4 matrix of juvenile fish size s, m, l, vl}

\item{inchannel_habitat}{The available tributary habitat in square meters}

\item{floodplain_habitat}{The available floodplain habitat in square meters}

\item{territory_size}{A length 4 array of juvenile fish territory requirements s, m, l, vl}

\item{up_to_size_class}{Which size class and under is moved to floodplain and
inchannel habitat (small = 1, medium = 2, large = 3, very large = 4)}
}
\description{
Allocates juvenile fish onto the floodplain, in-channel, and once
available habitat is full in the natal stream, assigns fish to out migrate
}
