% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_regional}
\alias{fill_regional}
\title{Filling Regional Habitat}
\source{
IP-117068
}
\usage{
fill_regional(
  juveniles,
  habitat,
  floodplain_habitat = NULL,
  territory_size = c(0.0498944803729701, 0.138941944739835, 0.471083652829798, 0),
  up_to_size_class = 3
)
}
\arguments{
\item{juveniles}{An n by 4 matrix of juvenile fish size s, m, l, vl}

\item{habitat}{The available habitat in square meters (in-channel for tributaries)}

\item{floodplain_habitat}{The available tributary floodplain habitat in square meters (NULL for bypasses and delta)}

\item{territory_size}{A length 4 array of juvenile fish territory requirements s, m, l, vl}

\item{up_to_size_class}{Which size class and under is moved to floodplain and
inchannel habitat (small = 1, medium = 2, large = 3, very large = 4)}
}
\description{
Allocates juvenile fish onto the floodplain, in-channel, and once
available habitat is full, assigns fish to out migrate. Function is used to route fish
on the mainstem, bypasses, and deltas.
}
