% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{get_growth_rates}
\alias{get_growth_rates}
\title{Get Growth Rates}
\usage{
get_growth_rates(
  temperature,
  prey_density = c("low", "med", "hi", "max"),
  floodplain = FALSE,
  transitions = fallRunDSM::bioenergetics_transitions
)
}
\arguments{
\item{temperature}{in Celsius to be used to lookup}

\item{prey_density}{one of "low", "med", "hi", or "max"}

\item{floodplain}{boolean indicating whether lookup is for floodplain}
}
\description{
Get Growth Rates
}
