% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_habitat.R
\name{get_habitat}
\alias{get_habitat}
\title{Get habitat Values}
\source{
IP-117068
}
\usage{
get_habitat(
  year,
  month,
  inchannel_habitat_fry,
  inchannel_habitat_juvenile,
  floodplain_habitat,
  sutter_habitat,
  yolo_habitat,
  delta_habitat
)
}
\arguments{
\item{year}{The current simulation year, 1-20}

\item{month}{The current simulation month, 1-8}

\item{inchannel_habitat_fry}{Variable describing fry inchannel rearing habitat}

\item{inchannel_habitat_juvenile}{Variable describing juvenile inchannel rearing habitat}

\item{floodplain_habitat}{Variable describing floodplain rearing habitat}

\item{sutter_habitat}{Variable describing sutter bypass rearing habitat}

\item{yolo_habitat}{Variable describing yolo bypass rearing habitat}

\item{north_delta_habitat}{Variable describing north delta rearing habitat}

\item{south_delta_habitat}{Variable describing south delta rearing habitat}
}
\description{
Returns the appropriate habitat for spawning and rearing based on habitat inputs and a given simulation year and month
}
\details{
See \code{\link{params}} for details on parameter sources
}
