% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_migratory_survival_rates}
\alias{get_migratory_survival_rates}
\title{Get Migratory Survival Rates}
\source{
IP-117068
}
\usage{
get_migratory_survival_rates(
  year,
  month,
  ..surv_juv_outmigration_sj_int = ..surv_juv_outmigration_sj_int,
  ..surv_juv_outmigration_sac_int_one = ..surv_juv_outmigration_sac_int_one,
 
    ..surv_juv_outmigration_sac_prop_diversions = ..surv_juv_outmigration_sac_prop_diversions,
 
    ..surv_juv_outmigration_sac_total_diversions = ..surv_juv_outmigration_sac_total_diversions,
  ..surv_juv_outmigration_sac_int_two = ..surv_juv_outmigration_sac_int_two
)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}
}
\description{
Calculates the juvenile out migration survival rates in all
regions for a month and year of the simulation
}
