% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Growth Transition Probability}
\source{
IP-117068
}
\usage{
growth(daily_growth_rate = 0.5, size_class_breaks = c(35, 42, 72, 110))
}
\arguments{
\item{daily_growth_rate}{The daily growth rate for inchannel (default value is 0.5 mm/day)}

\item{size_class_breaks}{The fork length size class breaks (default values are s = 35mm, m = 42mm, l = 72mm, vl = 110mm))}
}
\description{
Generates transition probability matrices for growth inchannel
}
