% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_rates}
\alias{growth_rates}
\alias{growth_rates_inchannel}
\alias{growth_rates_floodplain}
\title{Growth Rates}
\usage{
growth_rates_inchannel

growth_rates_floodplain
}
\description{
Growth rates determined by expert SIT judgment based on the following:
\itemize{
  \item \href{https://link.springer.com/article/10.1007/s10641-009-9473-8}{Limm, M.P., Marchetti, M.P.  (2009)}
  \item \href{https://www.tandfonline.com/doi/abs/10.1577/T08-112.1}{Rene E. Henery, Ted R. Sommer & Charles R. Goldman (2010)}
  \item \href{https://link.springer.com/article/10.1007/s10641-008-9367-1}{Jeffres, C.A., Opperman, J.J. & Moyle, P.B. (2008)}
}
}
\keyword{datasets}
