% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest_adults}
\alias{harvest_adults}
\title{Harvest Adults Function}
\usage{
harvest_adults(
  adult_df,
  year = year,
  hatchery_allocation = ..params$hatchery_allocation,
  ocean_harvest_percentage,
  tributary_harvest_percentage,
  restrict_harvest_to_hatchery = c(F, T),
  no_cohort_harvest_years,
  intelligent_crr_harvest = c(F, T)
)
}
\arguments{
\item{adult_df}{A data frame containing information about adult salmon.}

\item{year}{The target year for the harvest calculation.}

\item{hatchery_allocation}{The allocation of harvest to hatcheries.}

\item{ocean_harvest_percentage}{The percentage of harvest from the ocean, same for all tributaries.}

\item{tributary_harvest_percentage}{The percentage of harvest from tributaries, can vary by tributary.}

\item{restrict_harvest_to_hatchery}{A logical vector indicating whether to restrict harvest to hatcheries for each tributary.}

\item{no_cohort_harvest_years}{A vector of years where no cohort harvest is allowed.}

\item{intelligent_crr_harvest}{A logical vector indicating whether to apply intelligent harvest based on Coho Retention Rate (CRR).}
}
\value{
A numeric value representing the total adult salmon harvest.
}
\description{
This function calculates the total adult salmon harvest based on various parameters.
This function is called within the larger model.R file before stray rates and prespawn survival is applied.
}
