% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_allocation}
\alias{hatchery_allocation}
\title{Hatchery Allocation}
\format{
1 dimensional array [31 watersheds]
}
\source{
Various \href{https://s3-us-west-2.amazonaws.com/cvpia-reference-docs/AWP+Attachment+1.pdf}{CVPIA Science Integration Team: FY18 Decision Support Model activities and FY17 priorities Memorandum}

Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
hatchery_allocation
}
\description{
The proportion of hatchery fish spawning
}
\keyword{datasets}
