% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_releases_at_chipps}
\alias{hatchery_releases_at_chipps}
\title{Hatchery Release at Chipps}
\format{
A matrix with dimension 31 x 4 (watershed x size class)
}
\source{
Hatchery return data comes from the production targets defined in the
\href{https://swfsc-publications.fisheries.noaa.gov/publications/CR/2012/2012California.pdf}{California HSRG (Pages 65 - 95)}
}
\usage{
hatchery_releases_at_chipps
}
\description{
Hatchery release at Chipps by size class and hatchery
}
\keyword{datasets}
