% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_stray_betareg}
\alias{hatchery_stray_betareg}
\title{Hatchery Stray Model}
\format{
An object of class \code{betareg} of length 29.
}
\source{
Model from
 \href{https://afspubs.onlinelibrary.wiley.com/doi/10.1002/fsh.10267}{Sturrock et al 2019}
}
\usage{
hatchery_stray_betareg
}
\description{
A serialized version of the beta-regression model for hatchery origin fish straying
}
\keyword{datasets}
