% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_to_bay_distance}
\alias{hatchery_to_bay_distance}
\title{Hatchery to Chipps Island (rkm)}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
hatchery_to_bay_distance
}
\description{
Lookup for distances from hatchery to the Chipps Island, measured in river kilometers
}
\keyword{datasets}
