% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_month_dynamic.R
\name{juvenile_month_dynamic}
\alias{juvenile_month_dynamic}
\title{Juvenile Month Dynamics}
\usage{
juvenile_month_dynamic(
  hypothesis,
  fish,
  year = year,
  month = month,
  rearing_survival = rearing_survival,
  migratory_survival = migratory_survival,
  habitat = habitat,
  ..params = ..params,
  avg_ocean_transition_month = avg_ocean_transition_month,
  stochastic = stochastic,
  juveniles,
  natal_filling_fn = fallRunDSM::fill_natal,
  regional_filling_fn = fallRunDSM::fill_regional,
  movement_fn = NULL
)
}
\arguments{
\item{fish}{list object containing tracking matrices for juveniles, juveniles at chips, and adults in ocean}

\item{natal_filling_fn}{overwrite base filling function}

\item{regional_filling_fn}{overwrite base regional filling}

\item{movement_fn}{adds a additional movement function}
}
\description{
Juvenile Month Dynamics
}
\examples{
fish = list(juveniles = juveniles, north_delta_fish = north_delta_fish, south_delta_fish = south_delta_fish,
juveniles_at_chipps = juveniles_at_chipps,
adults_in_ocean = adults_in_ocean)
}
