% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-baseline-data.R
\name{load_baseline_data}
\alias{load_baseline_data}
\title{Load Baseline Data}
\usage{
load_baseline_data()
}
\description{
loads baseline data from DSMflow, DSMhabitat and DSMtemperature
}
\details{
More details for each item in the list available below.

\strong{Flow Inputs}
\itemize{
  \item freeport_flows - \link[DSMflow]{freeport_flow}
  \item vernalis_flows - \link[DSMflow]{vernalis_flow}
  \item stockton_flows - \link[DSMflow]{stockton_flow}
  \item CVP_exports - \link[DSMflow]{cvp_exports}
  \item SWP_exports - \link[DSMflow]{swp_exports}
  \item proportion_diverted - \link[DSMflow]{proportion_diverted}
  \item total_diverted - \link[DSMflow]{total_diverted}
  \item delta_proportion_diverted - \link[DSMflow]{delta_proportion_diverted}
  \item delta_total_diverted - \link[DSMflow]{delta_total_diverted}
  \item prop_pulse_flows - \link[DSMflow]{proportion_pulse_flows}
  \item prop_flow_natal - \link[DSMflow]{proportion_flow_natal}
  \item upper_sacramento_flows - \link[DSMflow]{upper_sacramento_flows}
  \item delta_inflow - \link[DSMflow]{delta_inflow}
  \item cc_gates_days_closed - \link[DSMflow]{delta_cross_channel_closed}
  \item cc_gates_prop_days_closed - \link[DSMflow]{delta_cross_channel_closed}
  \item proportion_flow_bypass - \link[DSMflow]{proportion_flow_bypasses}
  \item gates_overtopped - \link[DSMflow]{gates_overtopped}
}

\strong{Temperature Inputs}
\itemize{
  \item vernalis_temps - \link[DSMTemperature]{vernalis_temperature}
  \item prisoners_point_temps - \link[DSMTemperature]{prisoners_point_temperature}
  \item degree_days - \link[DSMTemperature]{degree_days}
  \item mean_egg_temp_effect - \link[DSMTemperature]{egg_temperature_effect}
  \item avg_temp - \link[DSMTemperature]{stream_temperature}
  \item avg_temp_delta - \link[DSMTemperature]{delta_temperature}
  \item migratory_temperature_proportion_over_20 - \link[DSMTemperature]{migratory_temperature_proportion_over_20}
}

\strong{Habitat Inputs}
\itemize{
  \item spawning_habitat - \link[DSMhabitat]{fr_spawn}
  \item inchannel_habitat_fry - \link[DSMhabitat]{fr_fry}
  \item inchannel_habitat_juvenile - \link[DSMhabitat]{fr_juv}
  \item floodplain_habitat - \link[DSMhabitat]{fr_fp}
  \item weeks_flooded - \link[DSMhabitat]{weeks_flooded}
  \item delta_habitat - \link[DSMhabitat]{delta_habitat}
  \item sutter_habitat - \link[DSMhabitat]{sutter_habitat}
  \item yolo_habitat - \link[DSMhabitat]{yolo_habitat}
  \item tisdale_bypass_watershed - \link[DSMhabitat]{tisdale_bypass_watershed}
  \item yolo_bypass_watershed - \link[DSMhabitat]{yolo_bypass_watershed}
  \item south_delta_routed_watersheds - \link[DSMhabitat]{south_delta_routed_watersheds}
  \item prop_high_predation - \link[DSMhabitat]{prop_high_predation}
  \item contact_points - \link[DSMhabitat]{contact_points}
  \item delta_contact_points - \link[DSMhabitat]{delta_contact_points}
  \item delta_prop_high_predation - \link[DSMhabitat]{delta_prop_high_predation}
  \item prob_strand_early - \link[DSMhabitat]{prob_strand_early}
  \item prob_strand_late - \link[DSMhabitat]{prob_strand_late}
  \item prob_nest_scoured - \link[DSMhabitat]{prob_nest_scoured}
}
}
