% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mass_by_size_class}
\alias{mass_by_size_class}
\title{Mass by Size Class}
\format{
1 dimensional array [4 size classes]
}
\source{
TODO
}
\usage{
mass_by_size_class
}
\description{
mass of fish by the size class
}
\details{
Mass by Size Class
}
\keyword{datasets}
