% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Juvenile Migration}
\source{
IP-117068
}
\usage{
migrate(migrants, migration_survival_rate)
}
\arguments{
\item{migrants}{An n by 4 matrix of juveniles (watersheds by size class)}

\item{migration_survival_rate}{The survival rates for s, m, l, and vl fish}
}
\description{
Stocastic application of survival rate to migrating juvenile fish
}
