% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{month_return_proportions}
\alias{month_return_proportions}
\title{Month Return Proportions}
\format{
1 dimensional array [31 watersheds]
}
\source{
TODO
}
\usage{
month_return_proportions
}
\description{
The proportion of spawning fish in Oct-Dec
}
\details{
Month Return Proportions
}
\keyword{datasets}
