% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{snow_globe_movement}
\alias{snow_globe_movement}
\alias{genetic_movement}
\title{Fry fish movement out of natal tributaries}
\usage{
snow_globe_movement(
  juveniles,
  freeport_flow,
  vernalis_flow,
  threshold = 1000,
  p_leave = 0.3,
  stochastic = FALSE
)

genetic_movement(juveniles, p_leave = 0.25, stochastic = FALSE)
}
\arguments{
\item{juveniles}{An n by 4 matrix of juvenile fish size s, m, l, vl}

\item{freeport_flow}{The flow at freeport in cubic meters per second}

\item{vernalis_flow}{The flow at vernalis in cubic meters per second}

\item{threshold}{the threshold combined flow for initiating fish movement out of natal tributaries}

\item{p_leave}{the proportion of juvenile fish leaving in response to high flows}

\item{stochastic}{when true the fish leaving are assigned randomly}
}
\description{
Sends fry Chinook salmon out of natal tributaries when they are at the fry stage
The \code{snow_globe_movement} function sends fish out of natal tributaries
           when the sum of flows at freeport and vernalis exceed a threshold
The \code{genetic_movement} function sends fish out of natal tributaries
          based on a user specified proportion
}
