% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_num}
\alias{pretty_num}
\title{Pretty Number}
\usage{
pretty_num(num, places = 2)
}
\arguments{
\item{num}{A number to be reformatted}

\item{places}{The number of places to round a number to, defaults to two}
}
\description{
Formats numbers for ease of reading
}
