% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prey_density}
\alias{prey_density}
\title{Prey Density
Vector defining prey density level ("low", "med", "hi", "max") by watershed}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 31 rows and 20 columns.
}
\usage{
prey_density
}
\description{
Prey Density
Vector defining prey density level ("low", "med", "hi", "max") by watershed
}
\keyword{datasets}
