% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prey_density_delta}
\alias{prey_density_delta}
\title{Prey Density Delta
Vector defining prey density level ("low", "med", "hi", "max") for North and South Delta}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 20 columns.
}
\usage{
prey_density_delta
}
\description{
Prey Density Delta
Vector defining prey density level ("low", "med", "hi", "max") for North and South Delta
}
\keyword{datasets}
