% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_hatchery}
\alias{proportion_hatchery}
\title{Proportion Hatchery}
\format{
1 dimensional array [31 watersheds]
}
\source{
TODO
}
\usage{
proportion_hatchery
}
\description{
Proportion Hatchery
}
\details{
Proportion Hatchery
}
\keyword{datasets}
