% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_movement.R
\name{pulse_movement}
\alias{pulse_movement}
\title{Juvenile Pulse Movement}
\source{
IP-117068
}
\usage{
pulse_movement(
  proportion_pulse,
  .intercept = fallRunDSM::params$.pulse_movement_intercept,
  .proportion_pulse = fallRunDSM::params$.pulse_movement_proportion_pulse,
  .medium = fallRunDSM::params$.pulse_movement_medium,
  .large = fallRunDSM::params$.pulse_movement_large,
  .vlarge = fallRunDSM::params$.pulse_movement_vlarge,
  .medium_pulse = fallRunDSM::params$.pulse_movement_medium_pulse,
  .large_pulse = fallRunDSM::params$.pulse_movement_large_pulse,
  .very_large_pulse = fallRunDSM::params$.pulse_movement_very_large_pulse
)
}
\arguments{
\item{proportion_pulse}{The proportion of flow that is a pulse, estimated
by the standard deviation of flow divided by the median flow for each month}

\item{.intercept}{Intercept}

\item{.proportion_pulse}{Coefficient for the proportion_pulse variable}

\item{.medium}{Parameter for medium sized fish}

\item{.large}{Parameter for large sized fish}

\item{.vlarge}{Parameter for very large sized fish}

\item{.medium_pulse}{Additional coefficient for proportion_pulse variable for medium size fish}

\item{.large_pulse}{Additional coefficient for proportion_pulse variable for large size fish}

\item{.very_large_pulse}{Additional coefficient for proportion_pulse variable for very large size fish}
}
\description{
Calculates the proportion of juveniles in each size class that
migrate due to pulse flow
}
