% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2r_adult_harvest_rate}
\alias{r2r_adult_harvest_rate}
\title{Adult Harvest Rate}
\format{
Named vector of length 31 corresponding to watersheds [31 watershed names]
}
\source{
average exploitation rate over the past ten years from the \href{https://www.pcouncil.org/documents/2022/03/2022-preseason-report-i.pdf/}{Ocean Salmon fishery preseason report 2022 table II-1}
}
\usage{
r2r_adult_harvest_rate
}
\description{
Adult harvest rate percentage by watershed
}
\keyword{datasets}
