% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{stray_returning_adults}
\alias{stray_returning_adults}
\title{Apply Straying to Returning Adults}
\usage{
stray_returning_adults(
  monthly_adult_returns,
  year,
  stochastic,
  month_return_proportions,
  wild,
  prop_flow_natal,
  south_delta_routed_watersheds,
  cc_gates_days_closed,
  prop_bay_trans = 0,
  prop_delta_trans = 0,
  type,
  total_releases,
  release_month,
  flows_oct_nov,
  flows_apr_may,
  monthly_mean_pdo,
  .adult_stray_intercept = fallRunDSM::params$.adult_stray_intercept,
  .adult_stray_wild = fallRunDSM::params$.adult_stray_wild,
  .adult_stray_natal_flow = fallRunDSM::params$.adult_stray_natal_flow,
  .adult_stray_cross_channel_gates_closed =
    fallRunDSM::params$.adult_stray_cross_channel_gates_closed,
  .adult_stray_prop_bay_trans = fallRunDSM::params$.adult_stray_prop_bay_trans,
  .adult_stray_prop_delta_trans = fallRunDSM::params$.adult_stray_prop_delta_trans
)
}
\description{
Apply Straying to Returning Adults
}
