% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_enroute}
\alias{surv_adult_enroute}
\title{Adult En Route Survival}
\source{
IP-117068
}
\usage{
surv_adult_enroute(
  migratory_temp,
  bypass_overtopped,
  adult_harvest,
  ..surv_adult_enroute_int = 3,
  .migratory_temp = -0.26,
  .bypass_overtopped = -0.019
)
}
\arguments{
\item{migratory_temp}{variable representing proportion of migratory corridor temperature above  20°C}

\item{bypass_overtopped}{Indicator for bypass overtopped}

\item{adult_harvest}{Adult harvest rate}

\item{..surv_adult_enroute_int}{Intercept}

\item{.migratory_temp}{Coefficient for \code{migratory_temp} variable}

\item{.bypass_overtopped}{Coefficient for \code{bypass_overtopped} variable}
}
\description{
Calculate adult survival en route to spawning grounds
}
