% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_prespawn}
\alias{surv_adult_prespawn}
\title{Adult Prespawn Survival}
\source{
IP-117068
}
\usage{
surv_adult_prespawn(
  deg_day,
  ..surv_adult_prespawn_int = winterRunDSM::params$..surv_adult_prespawn_int,
  .deg_day = winterRunDSM::params$.adult_prespawn_deg_day
)
}
\arguments{
\item{deg_day}{Variable describing average degree days}

\item{..surv_adult_prespawn_int}{Intercept}

\item{.deg_day}{Coefficient for \code{deg_day} variable}
}
\description{
Calculate the adult prespawn survival
}
\details{
See \code{\link{params}} for details on parameter sources
}
