% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_delta}
\alias{surv_juv_delta}
\title{Juvenile Delta Survival}
\source{
IP-117068
}
\usage{
surv_juv_delta(
  avg_temp,
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diverted,
  total_diverted,
  ..surv_juv_delta_int = 1.4,
  .avg_temp_thresh = -0.717,
  .high_predation = -0.122,
  ..surv_juv_delta_contact_points = 0.0358 * -0.189,
  .prop_diverted = -3.51,
  ..surv_juv_delta_total_diverted = 0.5 * -0.0021,
  .medium = 1.48,
  .large = 2.223
)
}
\arguments{
\item{avg_temp}{Variable representing average temperature in the delta}

\item{max_temp_thresh}{Variable representing the probability of exceeding the max temperature}

\item{avg_temp_thresh}{Variable representing the probability of exceeding the average temperature}

\item{high_predation}{Variable representing an indicator for high predation in delta}

\item{contact_points}{Variable representing the number of contact points in watershed}

\item{prop_diverted}{Variable representing the proportion of water diverted}

\item{total_diverted}{Variable representing the total diversions}

\item{..surv_juv_delta_int}{intercept}

\item{.avg_temp_thresh}{Coefficient for \code{avg_temp_thresh} variable}

\item{.high_predation}{Coefficient for \code{high_predation} variable}

\item{..surv_juv_delta_contact_points}{Coefficient for \code{contact_points} variable}

\item{.prop_diverted}{Coefficient for \code{prop_diversions} variable}

\item{..surv_juv_delta_total_diverted}{Coefficient for \code{total_diversions} variable}

\item{.medium}{parameter for medium sized fish}

\item{.large}{parameter for large sized fish}
}
\description{
Calculates the juvenile rearing survival in the deltas
}
