% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_delta}
\alias{surv_juv_outmigration_delta}
\title{Juvenile Delta Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_delta(
  prop_DCC_closed,
  hor_barr,
  freeport_flow,
  vernalis_flow,
  stockton_flow,
  vernalis_temperature,
  prisoners_point_temperature,
  CVP_exp,
  SWP_exp,
  trap_trans
)
}
\arguments{
\item{prop_DCC_closed}{proportion of days the Delta Cross Channel Gates are closed}

\item{hor_barr}{indicator if head of old river physical barrier in place}

\item{freeport_flow}{average daily discharge at Freeport in cubic meters per second}

\item{vernalis_flow}{average daily discharge at Vernalis in cubic meters per second}

\item{stockton_flow}{average daily discharge at Stockton in cubic meters per second}

\item{vernalis_temperature}{average daily temperature at Vernalis in °C}

\item{prisoners_point_temperature}{average daily temperature of the San Joaquin River at Prisoners Point °C}

\item{CVP_exp}{average daily exports Central Valley Project in cubic meters per second}

\item{SWP_exp}{average daily exports State Water Project in cubic meters per second}

\item{trap_trans}{proportion of smolts trapped at Vernalis and transported to Chips island}
}
\description{
Calculates the North and South Delta juvenile out migration survival
}
\details{
function returns proportion of fish from the Sacramento at Feeeport (northern_fish)
Mokelumne and Cosumnes (cosumnes_mokelumne_fish), Calaveras (calaveras_fish) and
San Joaquin tributaries from Vernalis (southern_fish) arriving alive at Chipps
Island in four size groups (35-42mm, 42-72mm, 72-110mm, >110mm).
Note that the models were fit to data that were >80 mm. Therefore, this does not
predict outside of the data so sizes <= 80mm are assumed to me 80mm long as
requested by Russ Perry.
}
\section{Parameters}{

All parameters were derived from Perry et al. (2018)
}

