% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_rear}
\alias{surv_juv_rear}
\title{Juvenile Rearing Survival}
\source{
IP-117068
}
\usage{
surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = 3.5,
  .avg_temp_thresh = -0.717,
  .high_predation = -0.122,
  ..surv_juv_rear_contact_points = -0.0068,
  ..surv_juv_rear_prop_diversions = -0.1755,
  ..surv_juv_rear_total_diversions = -5e-04,
  .stranded = -1.939,
  .medium = 1.48,
  .large = 2.223,
  .floodplain = 0.47
)
}
\arguments{
\item{max_temp_thresh}{variable representing probability of exceeding the max temperature threshold}

\item{avg_temp_thresh}{variable representing probability of exceeding the avg temperature threshold}

\item{high_predation}{variable representing indicator of high predation for a watershed}

\item{contact_points}{variable representing total number of contact points per watershed}

\item{prop_diversions}{variable representing proportion of water diverted}

\item{total_diversions}{variable representing total amount of water diverted}

\item{stranded}{variable representing stranding rate per watershed}

\item{weeks_flooded}{variable representing total weeks flooded per watershed}

\item{..surv_juv_rear_int}{Intercept}

\item{.avg_temp_thresh}{Coefficient for \code{avg_temp_thresh} variable}

\item{.high_predation}{Coefficient for \code{high_predation} variable}

\item{..surv_juv_rear_contact_points}{Coefficient for \code{contact_points} variable}

\item{..surv_juv_rear_prop_diversions}{Coefficient for \code{prop_diversions} variable}

\item{..surv_juv_rear_total_diversions}{Coefficient for \code{total_diversions} variable}

\item{.stranded}{Coefficient for \code{stranded} variable}

\item{.medium}{parameter for medium sized fish}

\item{.large}{parameter for large sized fish}

\item{.floodplain}{parameter for floodplain rearing benefit}
}
\description{
Calculates the juvenile rearing survival inchannel and on the floodplain
}
