% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_betas}
\alias{survival_betas}
\title{Survival Betas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 16 columns.
}
\usage{
survival_betas
}
\description{
calibration derived parameters used in the survival submodels.
}
\keyword{datasets}
