% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{temperature_movement}
\alias{temperature_movement}
\title{Temperature Movement}
\usage{
temperature_movement(juveniles, Month = 3, MeanT = 15, stochastic)
}
\description{
Move juvenile based on temperature metrics
}
