% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_seeds}
\alias{adult_seeds}
\title{Adult Seeds}
\format{
A matrix with dimension 31 x 30 (watershed x year)
}
\usage{
adult_seeds
}
\description{
adult fish for the initial 5 years of the simulations derived
from average escapement estimates from 2013 to 2017 (Azat 2019).
}
\keyword{datasets}
