% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{adult_stray}
\alias{adult_stray}
\title{Adult Straying}
\source{
IP-117068
}
\usage{
adult_stray(
  wild,
  natal_flow,
  south_delta_watershed,
  cross_channel_gates_closed,
  prop_bay_trans = 0,
  prop_delta_trans = 0,
  .intercept = fallRunDSM::params$.adult_stray_intercept,
  .wild = fallRunDSM::params$.adult_stray_wild,
  .natal_flow = fallRunDSM::params$.adult_stray_natal_flow,
  .cross_channel_gates_closed =
    fallRunDSM::params$.adult_stray_cross_channel_gates_closed,
  .prop_bay_trans = fallRunDSM::params$.adult_stray_prop_bay_trans,
  .prop_delta_trans = fallRunDSM::params$.adult_stray_prop_delta_trans
)
}
\arguments{
\item{wild}{Variable indicator of wild fish returning}

\item{natal_flow}{Variable describing proportion flows at tributary junctions coming from natal watershed in October}

\item{south_delta_watershed}{Variable indicator if watershed feeds into South Delta}

\item{cross_channel_gates_closed}{Variable describing number of days gates are closed for each month}

\item{prop_bay_trans}{Variable describing proportion transport to the bay}

\item{prop_delta_trans}{Variable describing proportion transport to the delta}

\item{.intercept}{Intercept}

\item{.wild}{Coefficient for \code{wild} variable}

\item{.natal_flow}{Coefficient for \code{natal_flow} variable}

\item{.cross_channel_gates_closed}{Coefficient for \code{cross_channel_gates_closed} variable}

\item{.prop_bay_trans}{Coefficient for \code{prop_bay_trans} variable}

\item{.prop_delta_trans}{Coefficient for \code{prop_delta_trans} variable}
}
\description{
Calculate the proportion of adults straying to non-natal streams to spawn
}
\details{
See \code{\link{params}} for details on parameter sources
}
