% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{adult_stray}
\alias{adult_stray}
\title{Adult Straying}
\source{
IP-117068
}
\usage{
adult_stray(
  wild,
  natal_flow,
  south_delta_watershed,
  cross_channel_gates_closed,
  prop_bay_trans = 0,
  prop_delta_trans = 0,
  .intercept = 3,
  .wild = -5.5,
  .natal_flow = -1.99,
  .cross_channel_gates_closed = -0.174,
  .prop_bay_trans = 2.09,
  .prop_delta_trans = 2.89
)
}
\arguments{
\item{wild}{Variable indicator of wild fish returning}

\item{natal_flow}{Variable describing proportion flows at tributary junctions coming from natal watershed in October}

\item{south_delta_watershed}{Variable indicator if watershed feeds into South Delta}

\item{cross_channel_gates_closed}{Variable describing number of days gates are closed for each month}

\item{prop_bay_trans}{Variable describing proportion transport to the bay}

\item{prop_delta_trans}{Variable describing proportion transport to the delta}

\item{.intercept}{Intercept, source: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District}

\item{.wild}{Coefficient for \code{wild} variable, source: Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}}

\item{.natal_flow}{Coefficient for \code{natal_flow} variable, source: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District}

\item{.cross_channel_gates_closed}{Coefficient for \code{cross_channel_gates_closed} variable, Source: Empirical model fit using  2008–2011 tagging data provided by East Bay Municipal Utility District.}

\item{.prop_bay_trans}{Coefficient for \code{prop_bay_trans} variable, source: Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}}

\item{.prop_delta_trans}{Coefficient for \code{prop_delta_trans} variable, source: Estimated with coded wire tag data 2010–2013 \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{(Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)}}
}
\description{
Calculate the proportion of adults straying to non-natal streams to spawn
}
