% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_channel_stray_rate}
\alias{cross_channel_stray_rate}
\title{Cross Channel Stray Rate}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
cross_channel_stray_rate
}
\description{
TODO
}
\keyword{datasets}
