% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diversity_group}
\alias{diversity_group}
\title{Diversity Groups}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
diversity_group
}
\description{
SIT defined diversity groups developed from the Central Valley Chinook Salmon and Steelhead Recovery Plan \href{https://archive.fisheries.noaa.gov/wcr/publications/recovery_planning/salmon_steelhead/domains/california_central_valley/cv_chin_stlhd_r_plan_fs_071614.pdf}{NOAA 2014}
}
\keyword{datasets}
