% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{gamma_MOM}
\alias{gamma_MOM}
\title{Method of Moments Gamma Parameter Estimation}
\usage{
gamma_MOM(mu, sigma)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}
}
\description{
Estimate parameters for gamma distribution using the method of moments
}
