% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_habitat.R
\name{get_habitat}
\alias{get_habitat}
\title{Get habitat Values}
\source{
IP-117068
}
\usage{
get_habitat(
  year,
  month,
  inchannel_habitat_fry,
  inchannel_habitat_juvenile,
  floodplain_habitat,
  sutter_habitat,
  yolo_habitat,
  delta_habitat
)
}
\arguments{
\item{year}{The current simulation year, 1-20}

\item{month}{The current simulation month, 1-8}

\item{inchannel_habitat_fry}{3 dimensional array [watersheds, months, years] representing fry inchannel habitat in square meters}

\item{inchannel_habitat_juvenile}{3 dimensional array [watersheds, months, years] representing juvenile inchannel habitat in square meters}

\item{floodplain_habitat}{3 dimensional array [watersheds, months, years] representing floodplain habitat in square meters}

\item{sutter_habitat}{2 dimensional array [months, years] representing sutter bypass habitat in square meters}

\item{yolo_habitat}{2 dimensional array [months, years] representing yolo bypass habitat in square meters}

\item{north_delta_habitat}{2 dimensional array [months, years] representing north delta habitat in square meters}

\item{south_delta_habitat}{2 dimensional array [months, years] representing south delta habitat in square meters}
}
\description{
Returns the appropriate habitat for spawning and rearing based on habitat inputs and a given simulation year and month
}
\details{
See \code{\link{params}} for details on parameter sources
}
