% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_migratory_survival}
\alias{get_migratory_survival}
\title{Get Migratory Survival Rates}
\source{
IP-117068
}
\usage{
get_migratory_survival(
  year,
  month,
  cc_gates_prop_days_closed,
  freeport_flows,
  vernalis_flows,
  stockton_flows,
  vernalis_temps,
  prisoners_point_temps,
  CVP_exports,
  SWP_exports,
  upper_sacramento_flows,
  delta_inflow,
  avg_temp_delta,
  avg_temp,
  delta_proportion_diverted,
  ..surv_juv_outmigration_sj_int,
  .surv_juv_outmigration_san_joaquin_medium,
  .surv_juv_outmigration_san_joaquin_large,
  min_survival_rate,
  stochastic
)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}

\item{cc_gates_prop_days_closed}{More details at \code{\link[DSMflow]{delta_cross_channel_closed}}}

\item{freeport_flows}{More details at \code{\link[DSMflow]{freeport_flow}}}

\item{vernalis_flows}{More details at \code{\link[DSMflow]{vernalis_flow}}}

\item{stockton_flows}{More details at \code{\link[DSMflow]{stockton_flow}}}

\item{vernalis_temps}{More details at \code{\link[DSMtemperature]{vernalis_temperature}}}

\item{prisoners_point_temps}{More details at \code{\link[DSMtemperature]{prisoners_point_temperature}}}

\item{CVP_exports}{More details at \code{\link[DSMflow]{cvp_exports}}}

\item{SWP_exports}{More details at \code{\link[DSMflow]{swp_exports}}}

\item{upper_sacramento_flows}{average monthly flows on the Upper Sacramento River in cubic meters per second, more details at \code{\link[DSMflow]{upper_sacramento_flows}}}

\item{delta_inflow}{Variable describing delta inflow in cubic meters per second, more details at \code{\link[DSMflow]{delta_inflow}}}

\item{avg_temp_delta}{Variable describing monthly mean temperature in celsius, more details at  \code{\link[DSMtempetature]{delta_temprature}}}

\item{delta_proportion_diverted}{Variable describing diversions from the delta in cubic meters per seccond, more details at \code{\link[DSMflow]{delta_proportion_diverted}}}

\item{..surv_juv_outmigration_sj_int}{Intercept for \code{\link{surv_juv_outmigration_san_joaquin}}}

\item{.surv_juv_outmigration_san_joaquin_medium}{Size related intercept for \code{\link{surv_juv_outmigration_san_joaquin}} medium sized fish}

\item{.surv_juv_outmigration_san_joaquin_large}{Size related intercept for \code{\link{surv_juv_outmigration_san_joaquin}} large sized fish}

\item{min_survival_rate}{estimated survival rate if temperature threshold is exceeded}

\item{stochastic}{\code{TRUE} \code{FALSE} value indicating if model is being run stochastically}
}
\description{
Calculates the juvenile out migration survival rates in all
regions for a month and year of the simulation
}
