% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_migratory_survival_rates}
\alias{get_migratory_survival_rates}
\title{Get Migratory Survival Rates}
\source{
IP-117068
}
\usage{
get_migratory_survival_rates(
  year,
  month,
  cc_gates_prop_days_closed = fallRunDSM::params$cc_gates_prop_days_closed,
  freeport_flows = fallRunDSM::params$freeport_flows,
  vernalis_flows = fallRunDSM::params$vernalis_flows,
  stockton_flows = fallRunDSM::params$stockton_flows,
  vernalis_temps = fallRunDSM::params$vernalis_temps,
  prisoners_point_temps = fallRunDSM::params$prisoners_point_temps,
  CVP_exports = fallRunDSM::params$CVP_exports,
  SWP_exports = fallRunDSM::params$SWP_exports,
  upper_sacramento_flows = fallRunDSM::params$upper_sacramento_flows,
  delta_inflow = fallRunDSM::params$delta_inflow,
  avg_temp_delta = fallRunDSM::params$avg_temp_delta,
  avg_temp = fallRunDSM::params$avg_temp,
  delta_proportion_diverted = fallRunDSM::params$delta_proportion_diverted,
 
    .surv_juv_outmigration_sac_delta_intercept_one = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_one,
 
    .surv_juv_outmigration_sac_delta_intercept_two = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_two,
 
    .surv_juv_outmigration_sac_delta_intercept_three = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_three,
 
    .surv_juv_outmigration_sac_delta_delta_flow = fallRunDSM::params$.surv_juv_outmigration_sac_delta_delta_flow,
 
    .surv_juv_outmigration_sac_delta_avg_temp = fallRunDSM::params$.surv_juv_outmigration_sac_delta_avg_temp,
 
    .surv_juv_outmigration_sac_delta_perc_diversions = fallRunDSM::params$.surv_juv_outmigration_sac_delta_perc_diversions,
 
    .surv_juv_outmigration_sac_delta_medium = fallRunDSM::params$.surv_juv_outmigration_sac_delta_medium,
 
    .surv_juv_outmigration_sac_delta_large = fallRunDSM::params$.surv_juv_outmigration_sac_delta_large,
  ..surv_juv_outmigration_sj_int = fallRunDSM::params$..surv_juv_outmigration_sj_int,
 
    ..surv_juv_outmigration_sac_int_one = fallRunDSM::params$..surv_juv_outmigration_sac_int_one,
 
    ..surv_juv_outmigration_sac_prop_diversions = fallRunDSM::params$..surv_juv_outmigration_sac_prop_diversions,
 
    ..surv_juv_outmigration_sac_total_diversions = fallRunDSM::params$..surv_juv_outmigration_sac_total_diversions,
 
    ..surv_juv_outmigration_sac_int_two = fallRunDSM::params$..surv_juv_outmigration_sac_int_two,
 
    .surv_juv_outmigration_san_joquin_medium = fallRunDSM::params$.surv_juv_outmigration_san_joquin_medium,
 
    .surv_juv_outmigration_san_joaquin_large = fallRunDSM::params$.surv_juv_outmigration_san_joaquin_large
)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}

\item{cc_gates_prop_days_closed}{More details at \code{\link[DSMflow]{delta_cross_channel_closed}}}

\item{freeport_flows}{More details at \code{\link[DSMflow]{freeport_flow}}}

\item{vernalis_flows}{More details at \code{\link[DSMflow]{vernalis_flow}}}

\item{stockton_flows}{More details at \code{\link[DSMflow]{stockton_flow}}}

\item{vernalis_temps}{More details at \code{\link[DSMtemperature]{vernalis_temperature}}}

\item{prisoners_point_temps}{More details at \code{\link[DSMtemperature]{prisoners_point_temperature}}}

\item{CVP_exports}{More details at \code{\link[DSMflow]{cvp_exports}}}

\item{SWP_exports}{More details at \code{\link[DSMflow]{swp_exports}}}

\item{..surv_juv_outmigration_sj_int}{Intercept for \code{\link{surv_juv_outmigration_san_joaquin}}}

\item{..surv_juv_outmigration_sac_int_one}{Intercept one for \code{\link{surv_juv_outmigration_sac_delta}}}

\item{..surv_juv_outmigration_sac_prop_diversions}{Coefficent for code{\link{surv_juv_outmigration_sac_delta}} \code{prop_diversion} variable}

\item{..surv_juv_outmigration_sac_total_diversions}{Coefficent for code{\link{surv_juv_outmigration_sac_delta}} \code{total_diversion} variable #TODO (do we use this variable?)}

\item{..surv_juv_outmigration_sac_int_two}{Intercept one for \code{\link{surv_juv_outmigration_sac_delta}}}

\item{.surv_juv_outmigration_san_joquin_medium}{parameter for \code{\link{surv_juv_outmigration_san_joaquin}} medium sized fish}

\item{.surv_juv_outmigration_san_joaquin_large}{parameter for \code{\link{surv_juv_outmigration_san_joaquin}} large sized fish}
}
\description{
Calculates the juvenile out migration survival rates in all
regions for a month and year of the simulation
}
