% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_migratory_survival_rates}
\alias{get_migratory_survival_rates}
\title{Get Migratory Survival Rates}
\source{
IP-117068
}
\usage{
get_migratory_survival_rates(year, month)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}
}
\description{
Calculates the juvenile out migration survival rates in all
regions for a month and year of the simulation
}
