% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_rearing_survival_rates}
\alias{get_rearing_survival_rates}
\title{Get Rearing Survival Rates}
\source{
IP-117068
}
\usage{
get_rearing_survival_rates(
  year,
  month,
  scenario,
  ..surv_juv_rear_int,
  ..surv_juv_rear_contact_points,
  ..surv_juv_rear_prop_diversions,
  ..surv_juv_rear_total_diversions,
  ..surv_juv_bypass_int,
  ..surv_juv_delta_int,
  ..surv_juv_delta_contact_points,
  ..surv_juv_delta_total_diverted
)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}

\item{scenario}{The current scenario}
}
\description{
Calculates the juvenile inchannel, floodplain, bypasses, and
deltas rearing survival rates for a month and year of the simulation
}
