% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_rearing_survival_rates}
\alias{get_rearing_survival_rates}
\title{Get Rearing Survival Rates}
\source{
IP-117068
}
\usage{
get_rearing_survival_rates(year, month, scenario)
}
\arguments{
\item{year}{The simulation year, 1-20}

\item{month}{The simulation month, 1-8}

\item{scenario}{The current scenario}
}
\description{
Calculates the juvenile inchannel, floodplain, bypasses, and
deltas rearing survival rates for a month and year of the simulation
}
