% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spawning_adults.R
\name{get_spawning_adults}
\alias{get_spawning_adults}
\title{Get Spawning Adults}
\source{
IP-117068
}
\usage{
get_spawning_adults(
  year,
  adults,
  hatch_adults,
  mode,
  month_return_proportions = fallRunDSM::params$month_return_proportions,
  prop_flow_natal,
  south_delta_routed_watersheds,
  cc_gates_days_closed,
  gates_overtopped,
  tisdale_bypass_watershed,
  yolo_bypass_watershed,
  migratory_temperature_proportion_over_20,
  ..surv_adult_enroute_int,
  .adult_stray_intercept,
  .adult_stray_wild,
  .adult_stray_natal_flow,
  .adult_stray_cross_channel_gates_closed,
  .adult_stray_prop_bay_trans,
  .adult_stray_prop_delta_trans,
  .adult_en_route_migratory_temp,
  .adult_en_route_bypass_overtopped,
  .adult_en_route_adult_harvest_rate
)
}
\arguments{
\item{year}{the year of simulation}

\item{adults}{potential spawning adults for each watershed (length = 31) values must be integer}

\item{hatch_adults}{total hatchery adults}

\item{month_return_proportions}{The proportion of fish returning for each month}

\item{migratory_temperature_proportion_over_20}{TODO}

\item{seeds}{a value meant to be inherited to determine if model is in seeding stage or not}
}
\description{
Get Spawning Adults
}
