% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spawning_adults.R
\name{get_spawning_adults}
\alias{get_spawning_adults}
\title{Get Spawning Adults}
\source{
IP-117068
}
\usage{
get_spawning_adults(year, adults, hatch_adults)
}
\arguments{
\item{year}{the year of simulation}

\item{adults}{potential spawning adults for each watershed (length = 31) values must be integer}
}
\description{
Get Spawning Adults
}
