% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/growth.R
\docType{data}
\name{growth_rates}
\alias{growth_rates}
\alias{growth}
\title{Growth Transistion Probability}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
IP-117068
}
\usage{
growth_rates

growth(daily_growth_rate = 0.5, size_class_breaks = c(35, 42, 72, 110))
}
\arguments{
\item{size_class_breaks}{The fork size class breaks}

\item{daily_rates}{The daily growth rate for inchannel (default value is 0.5 mm/day)}
}
\description{
Generates transition probability matrices for growth inchannel
}
\keyword{datasets}
