% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Growth Transition Probability}
\source{
IP-117068
}
\usage{
growth(daily_growth_rate = 0.5, size_class_breaks = c(35, 42, 72, 110))
}
\arguments{
\item{size_class_breaks}{The fork size class breaks}

\item{daily_rates}{The daily growth rate for inchannel (default value is 0.5 mm/day)}
}
\description{
Generates transition probability matrices for growth inchannel
}
