% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_floodplain}
\alias{growth_floodplain}
\title{Floodplain Growth Transition Probability}
\source{
IP-117068
}
\usage{
growth_floodplain(daily_rates = c(0.5, 1.06), weeks_flooded = 1:4)
}
\arguments{
\item{daily_rates}{The daily growth rate for inchannel (0.5 mm/day) and floodplain (1.06 mm/day)}

\item{weeks_flooded}{The number of weeks inundated, 1-4 weeks}
}
\description{
Generates transition probability matrices for growth on the
floodplain dependent on number of weeks inundated
}
