% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_release_proportion_bay}
\alias{hatchery_release_proportion_bay}
\title{Bay Proportion Hatchery Releases}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
hatchery_release_proportion_bay
}
\description{
The proportion of hatchery releases occurring at the Bay. Note that the complement
\code{1 - hatchery_release_proportion_bay} is the proportion of releases occurring in-river.
}
\keyword{datasets}
