% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_to_bay_distance}
\alias{hatchery_to_bay_distance}
\title{Hatchery to Chipps Island (rkm)}
\format{
Named vector of length five corresponding to hatchery locations [5 hatchery names]
}
\source{
Distance from hatchery to Chipps Island pulled from table 1 of
 \href{https://afspubs.onlinelibrary.wiley.com/doi/10.1002/fsh.10267}{Sturrock et al 2019}
}
\usage{
hatchery_to_bay_distance
}
\description{
Lookup table for distances from hatchery to the Chipps Island, measured in river kilometers
}
\keyword{datasets}
