% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{month_return_proportions}
\alias{month_return_proportions}
\title{Month Return Proportions}
\format{
Vector of length 3
}
\source{
Expert opinion from SIT members.
}
\usage{
month_return_proportions
}
\description{
The proportion of spawning fish in Oct-Dec
}
\keyword{datasets}
