% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{month_return_proportions}
\alias{month_return_proportions}
\title{Month Return Proportions}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
month_return_proportions
}
\description{
the proportion of spawning fish in Oct-Dec
}
\keyword{datasets}
