% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_entry_success.R
\name{ocean_entry_success}
\alias{ocean_entry_success}
\title{Ocean Entry Success}
\source{
IP-117068
}
\usage{
ocean_entry_success(
  migrants,
  month,
  avg_ocean_transition_month,
  .ocean_entry_success_length = fallRunDSM::params$.ocean_entry_success_length,
  ..ocean_entry_success_int = fallRunDSM::params$..ocean_entry_success_int,
  .ocean_entry_success_months = fallRunDSM::params$.ocean_entry_success_months
)
}
\arguments{
\item{migrants}{Variable representing the number of juveniles at golden gate bridge}

\item{month}{Variable representing the current simulation month}

\item{avg_ocean_transition_month}{Variable representing the average month juveniles transition to the ocean}

\item{.ocean_entry_success_length}{Variable representing the fork lengths for each size classes}

\item{..ocean_entry_success_int}{Intercept}

\item{.ocean_entry_success_months}{Coefficient for month variable}
}
\description{
Calculates the number of juveniles that survive entering the ocean
}
\details{
See \code{\link{params}} for details on parameter sources
}
