% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_entry_success.R
\name{ocean_entry_success}
\alias{ocean_entry_success}
\title{Ocean Entry Success}
\source{
IP-117068
}
\usage{
ocean_entry_success(
  migrants,
  month,
  avg_ocean_transition_month,
  length = c(-0.0897309864, -0.0709704348, -0.0208590732, 0.0732620916),
  betas = c(`intercept 1` = -0.5108849, `intercept 2` = 1.2, `intercept 3` = 1.2,
    `intercept 4` = 1.2, `intercept 5` = 1.2, `intercept 6` = -3.3233638, `intercept 7` =
    1.2, `intercept 8` = 1.2, `intercept 9` = 1.2, `intercept 10` = -3.2304288,
    `intercept 11` = 1.2, `intercept 12` = -3.4148335, `intercept 13` = 1.2, `intercept
    14` = 1.2, `intercept 15` = 1.2, `intercept 16` = 1.2, `intercept 17` = 1.2,
    `intercept 18` = -3.5, `intercept 19` = -3.5, `intercept 20` = -3.5, `intercept 21` =
    1.2, `intercept 22` = 1.2,      `intercept 23` = -1.308341, `intercept 24` = 1.2,
    `intercept 25` = -1.9841364, `intercept 26` = -1.9841364, `intercept 27` = 2.5000007,
    `intercept 28` = -3.5, `intercept 29` = -3, `intercept 30` = -0.9, `intercept 31` =
    1.2, months = 0.35)
)
}
\arguments{
\item{migrants}{The number of juveniles at golden gate bridge}

\item{month}{The current simulation month}

\item{avg_ocean_transition_month}{The average month juveniles transition to the ocean}

\item{length}{Fork lengths for each size classes. \href{https://dsm-docs.s3-us-west-2.amazonaws.com/Satterthwaite_et_al_2014.pdf}{Satterthwaite et al. (2014)}}

\item{betas}{Parameters estimated through calibration}
}
\description{
Calculates the number of juveniles that survive entering the ocean
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
and meta-analysis. The source for each parameter in this function are detailed below.
\itemize{
\item intercept 1-31: calibration estimate; varies by tributary
\item months: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/Satterthwaite_et_al_2014.pdf}{Satterthwaite et al. (2014)}
}
}

