% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocean_transition_month}
\alias{ocean_transition_month}
\title{Ocean Transition Month}
\source{
IP-117068
}
\usage{
ocean_transition_month()
}
\description{
Calculates the month juveniles transition to the ocean
}
