% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params}
\alias{params}
\title{Model Parameters from Calibration}
\format{
An object of class \code{list} of length 17.
}
\usage{
params
}
\description{
calibration derived parameters used in submodels.
}
\keyword{datasets}
