% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_hatchery}
\alias{proportion_hatchery}
\title{Proportion Hatchery}
\format{
1 dimensional array [31 watersheds]
}
\source{
Coded Wire Tag Analysis and data from chinookprod database

More details \href{https://cvpia-osc.github.io/fallRunDSM/articles/hatchery-analysis.html}{here}
}
\usage{
proportion_hatchery
}
\description{
Proportion of spawning fish that are from a hatchery.
}
\keyword{datasets}
