% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_hatchery}
\alias{proportion_hatchery}
\title{Proportion Hatchery}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
proportion_hatchery
}
\description{
TODO
}
\keyword{datasets}
