% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_movement.R
\name{pulse_movement}
\alias{pulse_movement}
\title{Juvenile Pulse Movement}
\source{
IP-117068
}
\usage{
pulse_movement(
  proportion_pulse,
  betas = c(intercept = -7.70744, pulse = 0.26579, medium = 1.66845, large = 0.5706,
    vlarge = -4.305, `medium pulse` = -0.25477, `large pulse` = -0.44778, `very large
    pulse` = 0.329)
)
}
\arguments{
\item{proportion_pulse}{The proportion of flow that is a pulse, estimated
by the standard deviation of flow divided by the median flow for each month}

\item{betas}{parameters estimated through calibration}
}
\description{
Calculates the proportion of juveniles in each size class that
migrate due to pulse flow
}
\section{Parameters}{

All parameters were derived from Empirical model fit to CVPIA Chinook salmon screw
trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River,
Sacramento River, Tuolumne River, Clear Creek 2008–2015.
}

