% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_bypass}
\alias{route_bypass}
\title{Route Bypass}
\source{
IP-117068
}
\usage{
route_bypass(bypass_fish, bypass_habitat, migration_survival_rate)
}
\arguments{
\item{bypass_fish}{An n by 4 matrix of juvenile fish by watershed and size class}

\item{bypass_habitat}{A vector of available habitat in square meters}

\item{migration_survival_rate}{The outmigration survival rate}
}
\description{
Determines if juveniles remain in the bypass or out migrate
}
