% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_regional}
\alias{route_regional}
\title{Route Regions}
\source{
IP-117068
}
\usage{
route_regional(
  month,
  year,
  migrants,
  inchannel_habitat,
  floodplain_habitat,
  prop_pulse_flows,
  migration_survival_rate,
  proportion_flow_bypass,
  detour = NULL,
  territory_size,
  hypothesis,
  stochastic
)
}
\arguments{
\item{month}{The current simulation month, 1-8}

\item{year}{The current simulation year, 1-20}

\item{migrants}{An n by 4 matrix of juvenile fish by watershed and size class}

\item{inchannel_habitat}{A vector of available habitat in square meters}

\item{floodplain_habitat}{A vector of available floodplain habitat in square meters}

\item{prop_pulse_flows}{The proportion of pulse flows}

\item{migration_survival_rate}{The outmigration survival rate}

\item{proportion_flow_bypass}{Variable describing the proportion of flows routed through the bypasses, more details at \code{\link[DSMflow]{proportion_flow_bypasses}}}

\item{detour}{Values can be 'sutter' or 'yolo' if some juveniles are detoured on to that bypass, otherwise NULL}

\item{territory_size}{Array of juvenile fish territory requirements for \code{\link{fill_regional}}}

\item{stochastic}{\code{TRUE} \code{FALSE} value indicating if model is being run stochastically}
}
\description{
Determines if juveniles stay in the region (Sections of Mainstem
Sacramento River or San Joaquin River) or out migrate during a simulated month
}
