% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_regional}
\alias{route_regional}
\title{Route Regions}
\source{
IP-117068
}
\usage{
route_regional(
  month,
  migrants,
  inchannel_habitat,
  floodplain_habitat,
  prop_pulse_flows,
  migration_survival_rate
)
}
\arguments{
\item{month}{The simulation month, 1-8}

\item{migrants}{An n by 4 matrix of juvenile fish by watershed and size class}

\item{inchannel_habitat}{A vector of available habitat in square meters}

\item{floodplain_habitat}{A vector of available floodplain habitat in square meters}

\item{prop_pulse_flows}{The proportion of pulse flows}

\item{migration_survival_rate}{The outmigration survival rate}
}
\description{
Determines if juveniles stay in the region (Sections of Mainstem
Sacramento River or San Joaquin River) or out migrate during a simulated month
}
