% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_south_delta}
\alias{route_south_delta}
\title{South Delta Routing}
\usage{
route_south_delta(
  freeport_flow,
  dcc_closed,
  month,
  mean_freeport_flow = 21546.19,
  sd_freeport_flow = 14375.9,
  .sss_int = 1.892235,
  .sss_freeport_discharge = 2.170375,
  .sss_upper_asymptote = 0.3512465,
  .dcc_intercept = -1.48962,
  .dcc_freeport_discharge = -1.248865,
  .gs_intercept = -2.948145,
  .gs_freeport_discharge = -2.911835,
  .gs_dcc_effect_on_routing = -0.554843,
  .gs_lower_asymptote = 0.2729845
)
}
\arguments{
\item{freeport_flow}{Monthly mean flow at freeport in cubic feet per second}

\item{dcc_closed}{Number of days the Delta Cross Channel gates are closed during the month}

\item{month}{Current simulation month as an integer for calculating number of days the Delta Cross Channel gates are open}

\item{mean_freeport_flow}{Mean of flow at freeport for standardizing discharge}

\item{sd_freeport_flow}{Standard Deviation of flow at freeport for standardizing discharge}

\item{.sss_int}{Intercept for Sutter and Steamboat Sloughs, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.sss_freeport_discharge}{Coefficient for \code{freeport_flow} for Sutter and Steamboat Sloughs, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.sss_upper_asymptote}{Parameter representing the upper asymptote for Sutter and Steamboat Sloughs, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.dcc_intercept}{Intercept for Delta Cross Channel, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.dcc_freeport_discharge}{Coefficient for \code{freeport_flow} for Delta Cross Channel Gates, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.gs_intercept}{Intercept for Georgiana Slough, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.gs_freeport_discharge}{Coefficient for \code{freeport_flow} for Georgiana Slough, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.gs_dcc_effect_on_routing}{Parameter representing the dcc effect on routing, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}

\item{.gs_lower_asymptote}{Parameter representing the lower asymptote for Georgiana Slough, source: \href{https://doi.org/10.1139/cjfas-2017-0310}{This submodel is adapted from Perry et al. (2018)}}
}
\description{
Routes juveniles through the South Delta
}
