% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn_success.R
\name{spawn_success}
\alias{spawn_success}
\title{Spawning Success}
\source{
IP-117068
}
\usage{
spawn_success(
  escapement,
  adult_prespawn_survival,
  egg_to_fry_survival,
  prob_scour,
  spawn_habitat,
  sex_ratio = 0.5,
  redd_size = 9.29,
  fecundity = 5522
)
}
\arguments{
\item{escapement}{The number of returning adults}

\item{adult_prespawn_survival}{The adult prespawn surival rate}

\item{egg_to_fry_survival}{The egg to fry survival rate}

\item{prob_scour}{The probability of nest scouring}

\item{spawn_habitat}{The available spawning habitat in square meters}

\item{sex_ratio}{The female to male spawning ratio, default 0.5}

\item{redd_size}{The size of redds including defensible space, default value 9.29 square meters}

\item{fecundity}{The number of eggs per female, default value 5522}
}
\description{
Calculates the annual reproductive success.
}
