% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straying_destinations}
\alias{straying_destinations}
\title{Straying Destnations}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 31 rows and 6 columns.
}
\usage{
straying_destinations
}
\description{
Given a hatchery this matrix returns as rows the proportion of fish to stray at perticular watersheds.
}
\keyword{datasets}
