% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_enroute}
\alias{surv_adult_enroute}
\title{Adult En Route Survival}
\source{
IP-117068
}
\usage{
surv_adult_enroute(
  migratory_temp,
  bypass_overtopped,
  adult_harvest,
  betas = c(intercept = 3, temperature = -0.26, overtop = -0.019)
)
}
\arguments{
\item{migratory_temp}{Proportion of migratory corridor temperature above  20°C}

\item{bypass_overtopped}{Indicator for bypass overtopped}

\item{adult_harvest}{Adult harvest rate (Estimated with Coded Wire Tag data 2012–2013 (Palmer-Zwahlen & Kormos 2015; Palmer-Zwahlen et al. 2018))}

\item{betas}{Parameters estimated from calibration}
}
\description{
Calculate adult survial en route to spawning grounds
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
or meta-analysis. The source for each parameter in this function are detailed below.


\itemize{
\item intercept: Calibration Estimate
\item temperature (Average daily temperature): \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{Schreck et al. (1994)}
\item overtop (Bypass overtopped parameter): Expert opinion Ted Sommer, California Department of Water Resources (tributaries above bypasses only)
\item gates open - Cross Channel Gates Parameter: Empirical model fit using  2008–2011 tagging data provided by East Bay Municipal Utility District.
}
}

