% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_enroute}
\alias{surv_adult_enroute}
\title{Adult En Route Survival}
\source{
IP-117068
}
\usage{
surv_adult_enroute(
  migratory_temp,
  bypass_overtopped,
  adult_harvest,
  ..surv_adult_enroute_int = 3,
  .migratory_temp = -0.26,
  .bypass_overtopped = -0.019
)
}
\arguments{
\item{migratory_temp}{variable representing proportion of migratory corridor temperature above  20°C}

\item{bypass_overtopped}{Indicator for bypass overtopped}

\item{adult_harvest}{Adult harvest rate (Estimated with Coded Wire Tag data 2012–2013 (Palmer-Zwahlen & Kormos 2015; Palmer-Zwahlen et al. 2018))}

\item{..surv_adult_enroute_int}{intercept, source: calibration}

\item{.migratory_temp}{coefficient for \code{migratory_temp} variable, source: \href{https://nrm.dfg.ca.gov/FileHandler.ashx?DocumentID=162355&usg= AOvVaw0VgMOwD7knFfSxRZy6k8RG}{Schreck et al. (1994)}}

\item{.bypass_overtopped}{coefficient for \code{bypass_overtopped} variable, source: Expert opinion Ted Sommer, California Department of Water Resources (tributaries above bypasses only)}
}
\description{
Calculate adult survival en route to spawning grounds
}
