% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_prespawn}
\alias{surv_adult_prespawn}
\title{Adult Prespawn Survival}
\source{
IP-117068
}
\usage{
surv_adult_prespawn(
  deg_day,
  betas = c(intercept = 3, `degree days` = -0.000669526)
)
}
\arguments{
\item{deg_day}{Average degree days}

\item{betas}{Parameters from calibration process}
}
\description{
Calculate the adult prespawn survival
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
or meta-analysis. The source for each parameter in this function are detailed below.


\itemize{
\item intercept: Calibration Estimate
}
}

