% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_prespawn}
\alias{surv_adult_prespawn}
\title{Adult Prespawn Survival}
\source{
IP-117068
}
\usage{
surv_adult_prespawn(
  deg_day,
  ..surv_adult_prespawn_int = 3,
  .deg_day = -0.000669526
)
}
\arguments{
\item{deg_day}{Variable describing average degree days}

\item{..surv_adult_prespawn_int}{Intercept, Source: Calibration Estimate}

\item{.deg_day}{Coefficient for \code{deg_day} variable, source Colvin et al. (2018)}
}
\description{
Calculate the adult prespawn survival
}
