% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_bypass}
\alias{surv_juv_bypass}
\title{Juvenile Bypass Survival}
\source{
IP-117068
}
\usage{
surv_juv_bypass(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  ..surv_juv_bypass_int = -3.5,
  .avg_temp_thresh = -0.717,
  .high_predation = -0.122,
  .medium = 1.48,
  .large = 2.223,
  .floodplain = 0.47
)
}
\arguments{
\item{max_temp_thresh}{Variable representing the probability of exceeding the max temp threshold}

\item{avg_temp_thresh}{Variable representing the probability of exceeding the average temperature}

\item{high_predation}{Variable representing an indicator for high predation in watershed}

\item{..surv_juv_bypass_int}{intercept}

\item{.avg_temp_thresh}{coefficient for \code{avg_temp_thresh} variable}

\item{.high_predation}{coefficient for \code{high_predation} variable}

\item{.medium}{parameter for medium sized fish}

\item{.large}{parameter for large sized fish}

\item{.floodplain}{parameter for floodplain rearing benefit}
}
\description{
Calculates the juvenile rearing survival in the bypasses
}
