% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_bypass}
\alias{surv_juv_bypass}
\title{Juvenile Bypass Survival}
\source{
IP-117068
}
\usage{
surv_juv_bypass(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  ..surv_juv_bypass_int = springRunDSM::params$..surv_juv_bypass_int,
  .avg_temp_thresh = springRunDSM::params$.surv_juv_bypass_avg_temp_thresh,
  .high_predation = springRunDSM::params$.surv_juv_bypass_high_predation,
  .medium = springRunDSM::params$.surv_juv_bypass_medium,
  .large = springRunDSM::params$.surv_juv_bypass_large,
  .floodplain = springRunDSM::params$.surv_juv_bypass_floodplain,
  min_survival_rate = springRunDSM::params$min_survival_rate
)
}
\arguments{
\item{max_temp_thresh}{Variable representing the probability of exceeding the max temp threshold}

\item{avg_temp_thresh}{Variable representing the probability of exceeding the average temperature}

\item{high_predation}{Variable representing an indicator for high predation in watershed}

\item{..surv_juv_bypass_int}{intercept, source: calibration}

\item{.avg_temp_thresh}{coefficient for avg_temp_thresh variable}

\item{.high_predation}{coefficient for high_predation variable}

\item{.medium}{size related intercept for medium sized fish}

\item{.large}{size related intercept for large sized fish}

\item{.floodplain}{Additional intercept for floodplain rearing benefit}

\item{min_survival_rate}{estimated survival rate if temperature threshold is exceeded}
}
\description{
Calculates the juvenile rearing survival in the bypasses
}
\details{
See \code{\link{params}} for details on parameter sources
}
