% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_delta}
\alias{surv_juv_delta}
\title{Juvenile Delta Survival}
\source{
IP-117068
}
\usage{
surv_juv_delta(
  avg_temp,
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diverted,
  total_diverted,
  ..surv_juv_delta_int = fallRunDSM::params$..surv_juv_delta_int,
  .avg_temp_thresh = fallRunDSM::params$.surv_juv_delta_avg_temp_thresh,
  .high_predation = fallRunDSM::params$.surv_juv_delta_high_predation,
  .surv_juv_delta_contact_points = fallRunDSM::params$.surv_juv_delta_contact_points,
  ..surv_juv_delta_contact_points = fallRunDSM::params$..surv_juv_delta_contact_points,
  .prop_diverted = fallRunDSM::params$.surv_juv_delta_prop_diverted,
  .surv_juv_delta_total_diverted = fallRunDSM::params$.surv_juv_delta_total_diverted,
  ..surv_juv_delta_total_diverted = fallRunDSM::params$..surv_juv_delta_total_diverted,
  .medium = fallRunDSM::params$.surv_juv_delta_medium,
  .large = fallRunDSM::params$.surv_juv_delta_large,
  min_survival_rate = fallRunDSM::params$min_survival_rate,
  stochastic
)
}
\arguments{
\item{avg_temp}{Variable representing average temperature in the delta}

\item{max_temp_thresh}{Variable representing the probability of exceeding the max temperature}

\item{avg_temp_thresh}{Variable representing the probability of exceeding the average temperature}

\item{high_predation}{Variable representing an indicator for high predation in delta}

\item{contact_points}{Variable representing the number of contact points in watershed}

\item{prop_diverted}{Variable representing the proportion of water diverted}

\item{total_diverted}{Variable representing the total diversions}

\item{..surv_juv_delta_int}{intercept, source: calibration}

\item{.avg_temp_thresh}{Coefficient for \code{avg_temp_thresh} variable}

\item{.high_predation}{Coefficient for \code{high_predation} variable}

\item{.surv_juv_delta_contact_points}{Coefficient for \code{contact_points} variable}

\item{..surv_juv_delta_contact_points}{Calibrated coefficient for \code{contact_points} variable}

\item{.prop_diverted}{Coefficient for \code{prop_diversions} variable}

\item{.surv_juv_delta_total_diverted}{Coefficient for \code{total_diversions} variable}

\item{..surv_juv_delta_total_diverted}{Calibrated coefficient for \code{total_diversions} variable}

\item{.medium}{Size related intercept for medium sized fish}

\item{.large}{Size related intercept for large sized fish}

\item{min_survival_rate}{Estimated survival rate if temperature threshold is exceeded}

\item{stochastic}{\code{TRUE} \code{FALSE} value indicating if model is being run stochastically}
}
\description{
Calculates the juvenile rearing survival in the deltas
}
\details{
See \code{\link{params}} for details on parameter sources
}
