% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_delta}
\alias{surv_juv_delta}
\title{Juvenile Delta Survival}
\source{
IP-117068
}
\usage{
surv_juv_delta(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diverted,
  total_diverted,
  betas = c(intercept = 1.4, `avg temp thresh` = -0.717, predation = -0.122, contact =
    0.0358 * -0.189, `prop diversions` = -3.51, `total diversions` = 0.5 * -0.0021,
    medium = 1.48, large = 2.223)
)
}
\arguments{
\item{max_temp_thresh}{The probability of exceeding the max temp threshold}

\item{avg_temp_thresh}{The probability of exceeding the average temperature}

\item{high_predation}{An indicator for high predation in delta}

\item{contact_points}{The number of contact points in watershed}

\item{betas}{The parameter estimates from calibration}

\item{prop_diversions}{The proportion of water diverted}

\item{total_diversions}{The total diversions}
}
\description{
Calculates the juvenile rearing survival in the deltas
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
and meta-analysis. The source for each parameter in this function are detailed below.
\itemize{
\item intercept: calibration estimate; varies by tributary
\item average temperature: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/marine_cech_water_temp_effects.pdf}{Marine and Chech (2004)}
\item predation \href{https://pubag.nal.usda.gov/catalog/512123}{Cavallo et al. (2012)}
\item contact points: calibration estimate
\item proportion diverted: calibration estimate
\item total diverted: calibration estimate
\item medium: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}
\item large" \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}
}
}

