% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_delta}
\alias{surv_juv_delta}
\title{Juvenile Delta Survival}
\source{
IP-117068
}
\usage{
surv_juv_delta(
  avg_temp,
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diverted,
  total_diverted,
  ..surv_juv_delta_int = 1.4,
  .avg_temp_thresh = -0.717,
  .high_predation = -0.122,
  ..surv_juv_delta_contact_points = 0.0358 * -0.189,
  .prop_diverted = -3.51,
  ..surv_juv_delta_total_diverted = 0.5 * -0.0021,
  .medium = 1.48,
  .large = 2.223
)
}
\arguments{
\item{avg_temp}{Variable representing average temperature in the delta}

\item{max_temp_thresh}{Variable representing the probability of exceeding the max temperature}

\item{avg_temp_thresh}{Variable representing the probability of exceeding the average temperature}

\item{high_predation}{Variable representing an indicator for high predation in delta}

\item{contact_points}{Variable representing the number of contact points in watershed}

\item{prop_diverted}{Variable representing the proportion of water diverted}

\item{total_diverted}{Variable representing the total diversions}

\item{..surv_juv_delta_int}{intercept, source: calibration}

\item{.avg_temp_thresh}{Coefficient for avg_temp_thresh variable, source: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/marine_cech_water_temp_effects.pdf}{Marine and Chech (2004)}}

\item{.high_predation}{Coefficient for high_predation variable, source: \href{https://pubag.nal.usda.gov/catalog/512123}{Cavallo et al. (2012)}}

\item{..surv_juv_delta_contact_points}{Coefficient for contact_points variable, source: calibration}

\item{.prop_diverted}{Coefficient for prop_diversions variable, source: Newman and Brandes (2010)}

\item{..surv_juv_delta_total_diverted}{Coefficient for total_diversions variable, source: calibration}

\item{.medium}{parameter for medium sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.large}{parameter for large sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}
}
\description{
Calculates the juvenile rearing survival in the deltas
}
