% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac}
\alias{surv_juv_outmigration_sac}
\title{Juvenile Mainstem Sacramento Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac(
  flow_cms,
  avg_temp,
  total_diversions,
  prop_diversions,
  betas = c(`intercept 1` = 2.5, flow = 0.0092, `proportion diversion` = -3.51 * 0.05,
    `total diversion` = -0.0021 * 0.215, `intercept 2` = 0.3, `average temperature` =
    0.554, `model weight` = 0.5, medium = 1.48, large = 2.223)
)
}
\arguments{
\item{flow_cms}{Upper Sacramento River flow in cubic meters per second}

\item{avg_temp}{Monthly mean temperature in celsius}

\item{total_diversions}{Monthly mean total diversions in cubic feet per second}

\item{prop_diversions}{Monthly mean proportion diverted}

\item{betas}{Parameter estimates from calibration}
}
\description{
Calculates the Mainstem Sacramento juvenile out migration survival
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
and meta-analysis. The source for each parameter in this function are detailed below.
\itemize{
\item intercept 1 & 2: calibration estimate; varies by tributary
\item flow: emperical model fit to the 2014 late-fall-run Chinook salmon tag release data
\item proportion diverted: calibration estimate
\item total diverted: calibration estimate
\item average temperature: emperical model fit to the 2014 late-fall-run Chinook salmon tag release data
\item model weight:
\item medium: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/perry_2010.pdf}{Perry (2010)}
\item large" \href{https://dsm-docs.s3-us-west-2.amazonaws.com/perry_2010.pdf}{Perry (2010)}
}
}

