% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac}
\alias{surv_juv_outmigration_sac}
\title{Juvenile Mainstem Sacramento Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac(flow_cms)
}
\arguments{
\item{flow_cms}{Variable representing upper Sacramento River flow in cubic meters per second}
}
\description{
Calculates the Mainstem Sacramento juvenile out migration survival
}
