% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac_delta}
\alias{surv_juv_outmigration_sac_delta}
\title{Juvenile Delta Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac_delta(
  delta_flow,
  avg_temp,
  perc_diversions,
  .intercept_one = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_one,
  .intercept_two = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_two,
 
    .intercept_three = fallRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_three,
  .delta_flow = fallRunDSM::params$.surv_juv_outmigration_sac_delta_delta_flow,
  .avg_temp = fallRunDSM::params$.surv_juv_outmigration_sac_delta_avg_temp,
 
    .perc_diversions = fallRunDSM::params$.surv_juv_outmigration_sac_delta_perc_diversions,
  .medium = fallRunDSM::params$.surv_juv_outmigration_sac_delta_medium,
  .large = fallRunDSM::params$.surv_juv_outmigration_sac_delta_large
)
}
\arguments{
\item{delta_flow}{Variable describing delta inflow in cubic meters per second}

\item{avg_temp}{Variable describing monthly mean temperature in celsius}

\item{perc_diversions}{Variable describing monthly mean percent diverted}

\item{.intercept_one}{Intercept for model one}

\item{.intercept_two}{Intercept for model two}

\item{.intercept_three}{Intercept for model three}

\item{.delta_flow}{Coefficient for delta_flow variable}

\item{.avg_temp}{Coefficient for avg_temp variable}

\item{.perc_diversions}{Coefficient for perc_diversions variable}

\item{.medium}{parameter for medium sized fish}

\item{.large}{parameter for large sized fish}
}
\description{
Calculates the Sacramento Delta juvenile out migration survival
}
