% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac_delta}
\alias{surv_juv_outmigration_sac_delta}
\title{Juvenile Delta Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac_delta(
  delta_flow,
  avg_temp,
  perc_diversions,
  .intercept_one = winterRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_one,
  .intercept_two = winterRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_two,
 
    .intercept_three = winterRunDSM::params$.surv_juv_outmigration_sac_delta_intercept_three,
  .delta_flow = winterRunDSM::params$.surv_juv_outmigration_sac_delta_delta_flow,
  .avg_temp = winterRunDSM::params$.surv_juv_outmigration_sac_delta_avg_temp,
 
    .perc_diversions = winterRunDSM::params$.surv_juv_outmigration_sac_delta_perc_diversions,
  .medium = winterRunDSM::params$.surv_juv_outmigration_sac_delta_medium,
  .large = winterRunDSM::params$.surv_juv_outmigration_sac_delta_large,
  model_weights = winterRunDSM::params$surv_juv_outmigration_sac_delta_model_weights
)
}
\arguments{
\item{delta_flow}{Variable describing delta inflow in cubic meters per second}

\item{avg_temp}{Variable describing monthly mean temperature in celsius}

\item{perc_diversions}{Variable describing monthly mean percent diverted}

\item{.intercept_one}{Intercept for model one}

\item{.intercept_two}{Intercept for model two}

\item{.intercept_three}{Intercept for model three}

\item{.delta_flow}{Coefficient for delta_flow variable}

\item{.avg_temp}{Coefficient for avg_temp variable}

\item{.perc_diversions}{Coefficient for perc_diversions variable}

\item{.medium}{Size related intercept for medium sized fish}

\item{.large}{Size related intercept for large sized fish}

\item{model_weights}{weights for competing models}
}
\description{
Calculates the Sacramento Delta juvenile out migration survival
}
\details{
See \code{\link{params}} for details on parameter sources
}
