% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac_delta}
\alias{surv_juv_outmigration_sac_delta}
\title{Juvenile Delta Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac_delta(
  delta_flow,
  avg_temp,
  perc_diversions,
  betas = c(`intercept 1` = -3.5, `intercept 2` = 0.3, `intercept 3` = -3.5, flow =
    0.0013, `average temperature` = 0.386, `percent diversions` = -0.033, medium = 1.48,
    large = 2.223)
)
}
\arguments{
\item{avg_temp}{Monthly mean temperature in celsius}

\item{perc_diversions}{Monthly mean percent diverted}

\item{betas}{Parameter estimates from calibration}

\item{flow_cms}{Delta inflow in cubic meters per second}
}
\description{
Calculates the Sacramento Delta juvenile out migration survival
}
