% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_sac_delta}
\alias{surv_juv_outmigration_sac_delta}
\title{Juvenile Delta Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_sac_delta(
  delta_flow,
  avg_temp,
  perc_diversions,
  .intercept_one = -3.5,
  .intercept_two = 0.3,
  .intercept_three = -3.5,
  .delta_flow = 0.0013,
  .avg_temp = 0.386,
  .perc_diversions = -0.033,
  .medium = 1.48,
  .large = 2.223
)
}
\arguments{
\item{delta_flow}{Variable describing delta inflow in cubic meters per second}

\item{avg_temp}{Variable describing monthly mean temperature in celsius}

\item{perc_diversions}{Variable describing monthly mean percent diverted}

\item{.intercept_one}{Intercept for model one, source: TODO}

\item{.intercept_two}{Intercept for model two, source: TODO}

\item{.intercept_three}{Intercept for model three, source: TODO}

\item{.delta_flow}{Coefficient for delta_flow variable, source: TODO}

\item{.avg_temp}{Coefficient for avg_temp variable, source: TODO}

\item{.perc_diversions}{Coefficient for perc_diversions variable, source: TODO}

\item{.medium}{parameter for medium sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.large}{parameter for large sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}
}
\description{
Calculates the Sacramento Delta juvenile out migration survival
}
