% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_san_joaquin}
\alias{surv_juv_outmigration_san_joaquin}
\title{Juvenile San Joaquin Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_san_joaquin(
  ..surv_juv_outmigration_sj_int = winterRunDSM::params$..surv_juv_outmigration_sj_int,
  .medium = winterRunDSM::params$.surv_juv_outmigration_san_joaquin_medium,
  .large = winterRunDSM::params$.surv_juv_outmigration_san_joaquin_large
)
}
\arguments{
\item{..surv_juv_outmigration_sj_int}{Intercept}

\item{.medium}{Size related intercept for medium sized fish}

\item{.large}{Size related intercept for large sized fish}
}
\description{
Calculates the San Joaquin River juvenile out migration survival
}
\details{
See \code{\link{params}} for details on parameter sources
}
