% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_san_joaquin}
\alias{surv_juv_outmigration_san_joaquin}
\title{Juvenile San Joaquin Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_san_joaquin(
  betas = c(intercept = -3.5, medium = 1.48, large = 2.223)
)
}
\arguments{
\item{betas}{The parameters estimated through calibration}
}
\description{
Calculates the San Joaquin River juvenile out migration survival
}
\section{Parameters}{

Parameters from the model are obtained from either literature, calibration, export elicitation,
and meta-analysis. The source for each parameter in this function are detailed below.
\itemize{
\item intercept: calibration estimate
\item medium: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/perry_2010.pdf}{Perry (2010)}
\item large" \href{https://dsm-docs.s3-us-west-2.amazonaws.com/perry_2010.pdf}{Perry (2010)}
}
}

