% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_san_joaquin}
\alias{surv_juv_outmigration_san_joaquin}
\title{Juvenile San Joaquin Outmigration Survival}
\source{
IP-117068
}
\usage{
surv_juv_outmigration_san_joaquin(
  ..surv_juv_outmigration_sj_int = -3.5,
  .medium = 1.48,
  .large = 2.223
)
}
\arguments{
\item{..surv_juv_outmigration_sj_int}{intercept, source: calibration}

\item{.medium}{parameter for medium sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.large}{parameter for large sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}
}
\description{
Calculates the San Joaquin River juvenile out migration survival
}
