% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_rear}
\alias{surv_juv_rear}
\title{Juvenile Rearing Survival}
\source{
IP-117068
}
\usage{
surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = fallRunDSM::params$..surv_juv_rear_int,
  .avg_temp_thresh = fallRunDSM::params$.surv_juv_rear_avg_temp_thresh,
  .high_predation = fallRunDSM::params$.surv_juv_rear_high_predation,
  ..surv_juv_rear_contact_points = fallRunDSM::params$..surv_juv_rear_contact_points,
  ..surv_juv_rear_prop_diversions = fallRunDSM::params$..surv_juv_rear_prop_diversions,
 
    ..surv_juv_rear_total_diversions = fallRunDSM::params$..surv_juv_rear_total_diversions,
  .stranded = fallRunDSM::params$.surv_juv_rear_stranded,
  .medium = fallRunDSM::params$.surv_juv_rear_medium,
  .large = fallRunDSM::params$.surv_juv_rear_large,
  .floodplain = fallRunDSM::params$.surv_juv_rear_floodplain,
  min_survival_rate
)
}
\arguments{
\item{max_temp_thresh}{variable representing probability of exceeding the max temperature threshold}

\item{avg_temp_thresh}{variable representing probability of exceeding the avg temperature threshold}

\item{high_predation}{variable representing indicator of high predation for a watershed}

\item{contact_points}{variable representing total number of contact points per watershed}

\item{prop_diversions}{variable representing proportion of water diverted}

\item{total_diversions}{variable representing total amount of water diverted}

\item{stranded}{variable representing stranding rate per watershed}

\item{weeks_flooded}{variable representing total weeks flooded per watershed}

\item{..surv_juv_rear_int}{intercept, source: calibration (varies by tributary)}

\item{.avg_temp_thresh}{coefficient for avg_temp_thresh variable}

\item{.high_predation}{coefficient for high_predation variable, source: \href{https://pubag.nal.usda.gov/catalog/512123}{Cavallo et al. (2012)}}

\item{..surv_juv_rear_contact_points}{coefficient for contact_points variable, source: calibration}

\item{..surv_juv_rear_prop_diversions}{coefficient for prop_diversions variable, source: calibration}

\item{..surv_juv_rear_total_diversions}{coefficient for total_diversions variable, source: calibration}

\item{.stranded}{coefficient for stranded variable, source: \href{#}{USFWS (2006) and CDWR (2006)}}

\item{.medium}{parameter for medium sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.large}{parameter for large sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.floodplain}{parameter for floodplain rearing benefit, source: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/SOMMER_T-SDWA+180+Floodplain+rearing+of+juvenile+chinook+salmon+evidence+of+enhanced+growth+and+survival+.pdf}{Sommer et al. (2001)}}

\item{min_survival_rate}{estimated survival rate if temperature threshold is exceeded, source: expert opinion}
}
\description{
Calculates the juvenile rearing survival inchannel and on the floodplain
}
