% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_rear}
\alias{surv_juv_rear}
\title{Juvenile Rearing Survival}
\source{
IP-117068
}
\usage{
surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = springRunDSM::params$..surv_juv_rear_int,
  .avg_temp_thresh = springRunDSM::params$.surv_juv_rear_avg_temp_thresh,
  .high_predation = springRunDSM::params$.surv_juv_rear_high_predation,
  .surv_juv_rear_contact_points = springRunDSM::params$.surv_juv_rear_contact_points,
  ..surv_juv_rear_contact_points = springRunDSM::params$..surv_juv_rear_contact_points,
  .surv_juv_rear_prop_diversions = springRunDSM::params$.surv_juv_rear_prop_diversions,
 
    ..surv_juv_rear_prop_diversions = springRunDSM::params$..surv_juv_rear_prop_diversions,
 
    .surv_juv_rear_total_diversions = springRunDSM::params$.surv_juv_rear_total_diversions,
 
    ..surv_juv_rear_total_diversions = springRunDSM::params$..surv_juv_rear_total_diversions,
  .stranded = springRunDSM::params$.surv_juv_rear_stranded,
  .medium = springRunDSM::params$.surv_juv_rear_medium,
  .large = springRunDSM::params$.surv_juv_rear_large,
  .floodplain = springRunDSM::params$.surv_juv_rear_floodplain,
  min_survival_rate = springRunDSM::params$min_survival_rate
)
}
\arguments{
\item{max_temp_thresh}{variable representing probability of exceeding the max temperature threshold}

\item{avg_temp_thresh}{variable representing probability of exceeding the avg temperature threshold}

\item{high_predation}{variable representing indicator of high predation for a watershed}

\item{contact_points}{variable representing total number of contact points per watershed}

\item{prop_diversions}{variable representing proportion of water diverted}

\item{total_diversions}{variable representing total amount of water diverted}

\item{stranded}{variable representing stranding rate per watershed}

\item{weeks_flooded}{variable representing total weeks flooded per watershed}

\item{..surv_juv_rear_int}{intercept, source: calibration (varies by tributary)}

\item{.avg_temp_thresh}{coefficient for avg_temp_thresh variable}

\item{.high_predation}{coefficient for high_predation variable}

\item{.surv_juv_rear_contact_points}{coefficient for contact_points variable}

\item{..surv_juv_rear_contact_points}{calibrated coefficient for contact_points variable}

\item{.surv_juv_rear_prop_diversions}{coefficient for prop_diversions variable}

\item{..surv_juv_rear_prop_diversions}{calibrated coefficient for prop_diversions variable}

\item{.surv_juv_rear_total_diversions}{coefficient for total_diversions variable}

\item{..surv_juv_rear_total_diversions}{calibrated coefficient for total_diversions variable}

\item{.stranded}{coefficient for stranded variable}

\item{.medium}{size related intercept for medium sized fish}

\item{.large}{size related intercept for large sized fish}

\item{.floodplain}{Additional intercept for floodplain rearing benefit}

\item{min_survival_rate}{estimated survival rate if temperature threshold is exceeded, source: expert opinion}
}
\description{
Calculates the juvenile rearing survival inchannel and on the floodplain
}
\details{
See \code{\link{params}} for details on parameter sources
}
