% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_rear}
\alias{surv_juv_rear}
\title{Juvenile Rearing Survival}
\source{
IP-117068
}
\usage{
surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = fallRunDSM::params$..surv_juv_rear_int,
  .avg_temp_thresh = fallRunDSM::params$.surv_juv_rear_avg_temp_thresh,
  .high_predation = fallRunDSM::params$.surv_juv_rear_high_predation,
  .surv_juv_rear_contact_points = fallRunDSM::params$.surv_juv_rear_contact_points,
  ..surv_juv_rear_contact_points = fallRunDSM::params$..surv_juv_rear_contact_points,
  .surv_juv_rear_prop_diversions = fallRunDSM::params$.surv_juv_rear_prop_diversions,
  ..surv_juv_rear_prop_diversions = fallRunDSM::params$..surv_juv_rear_prop_diversions,
  .surv_juv_rear_total_diversions = fallRunDSM::params$.surv_juv_rear_total_diversions,
  ..surv_juv_rear_total_diversions = fallRunDSM::params$..surv_juv_rear_total_diversions,
  .stranded = fallRunDSM::params$.surv_juv_rear_stranded,
  .medium = fallRunDSM::params$.surv_juv_rear_medium,
  .large = fallRunDSM::params$.surv_juv_rear_large,
  .floodplain = fallRunDSM::params$.surv_juv_rear_floodplain,
  min_survival_rate = fallRunDSM::params$min_survival_rate,
  stochastic
)
}
\arguments{
\item{max_temp_thresh}{Variable representing probability of exceeding the max temperature threshold}

\item{avg_temp_thresh}{Variable representing probability of exceeding the avg temperature threshold}

\item{high_predation}{Variable representing indicator of high predation for a watershed}

\item{contact_points}{Variable representing total number of contact points per watershed}

\item{prop_diversions}{Variable representing proportion of water diverted}

\item{total_diversions}{Variable representing total amount of water diverted}

\item{stranded}{Variable representing stranding rate per watershed}

\item{weeks_flooded}{Variable representing total weeks flooded per watershed}

\item{..surv_juv_rear_int}{Intercept, source: calibration (varies by tributary)}

\item{.avg_temp_thresh}{Coefficient for \code{avg_temp_thresh} variable}

\item{.high_predation}{Coefficient for \code{high_predation} variable}

\item{.surv_juv_rear_contact_points}{Coefficient for \code{contact_points} variable}

\item{..surv_juv_rear_contact_points}{Calibrated coefficient for \code{contact_points} variable}

\item{.surv_juv_rear_prop_diversions}{Coefficient for \code{prop_diversions} variable}

\item{..surv_juv_rear_prop_diversions}{Calibrated coefficient for \code{prop_diversions} variable}

\item{.surv_juv_rear_total_diversions}{Coefficient for \code{total_diversions} variable}

\item{..surv_juv_rear_total_diversions}{Calibrated coefficient for \code{total_diversions} variable}

\item{.stranded}{Coefficient for \code{stranded} variable}

\item{.medium}{Size related intercept for medium sized fish}

\item{.large}{Size related intercept for large sized fish}

\item{.floodplain}{Additional intercept for floodplain rearing benefit}

\item{min_survival_rate}{Estimated survival rate if temperature threshold is exceeded}

\item{stochastic}{\code{TRUE} \code{FALSE} value indicating if model is being run stochastically}
}
\description{
Calculates the juvenile rearing survival inchannel and on the floodplain
}
\details{
See \code{\link{params}} for details on parameter sources
}
