% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_rear}
\alias{surv_juv_rear}
\title{Juvenile Rearing Survival}
\source{
IP-117068
}
\usage{
surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = 3.5,
  .avg_temp_thresh = -0.717,
  .high_predation = -0.122,
  ..surv_juv_rear_contact_points = -0.0068,
  ..surv_juv_rear_prop_diversions = -0.1755,
  ..surv_juv_rear_total_diversions = -5e-04,
  .stranded = -1.939,
  .medium = 1.48,
  .large = 2.223,
  .floodplain = 0.47
)
}
\arguments{
\item{max_temp_thresh}{variable representing probability of exceeding the max temperature threshold}

\item{avg_temp_thresh}{variable representing probability of exceeding the avg temperature threshold}

\item{high_predation}{variable representing indicator of high predation for a watershed}

\item{contact_points}{variable representing total number of contact points per watershed}

\item{prop_diversions}{variable representing proportion of water diverted}

\item{total_diversions}{variable representing total amount of water diverted}

\item{stranded}{variable representing stranding rate per watershed}

\item{weeks_flooded}{variable representing total weeks flooded per watershed}

\item{..surv_juv_rear_int}{intercept, source: calibration (varies by tributary)}

\item{.avg_temp_thresh}{coefficient for avg_temp_thresh variable, source: \href{https://www.noaa.gov/sites/default/files/atoms/files/07354626766.pdf}{Marine and Chech (2004)}}

\item{.high_predation}{coefficient for high_predation variable, source: \href{https://pubag.nal.usda.gov/catalog/512123}{Cavallo et al. (2012)}}

\item{..surv_juv_rear_contact_points}{coefficient for contact_points variable, source: calibration}

\item{..surv_juv_rear_prop_diversions}{coefficient for prop_diversions variable, source: calibration}

\item{..surv_juv_rear_total_diversions}{coefficient for total_diversions variable, source: calibration}

\item{.stranded}{coefficient for stranded variable, source: \href{#}{USFWS (2006) and CDWR (2006)}}

\item{.medium}{parameter for medium sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.large}{parameter for large sized fish, source: \href{https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1577/M02-161.1}{Connor et al. (2004)}}

\item{.floodplain}{parameter for floodplain rearing benefit, source: \href{https://dsm-docs.s3-us-west-2.amazonaws.com/SOMMER_T-SDWA+180+Floodplain+rearing+of+juvenile+chinook+salmon+evidence+of+enhanced+growth+and+survival+.pdf}{Sommer et al. (2001)}}
}
\description{
Calculates the juvenile rearing survival inchannel and on the floodplain
}
